requirements.txt
astroid==2.13.2
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
dill==0.3.6
exceptiongroup==1.1.0
Faker==16.3.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
isort==5.11.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
outcome==1.2.0
packaging==23.0
platformdirs==2.6.2
pluggy==1.0.0
pylint==2.15.10
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
PyYAML==6.0
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tomlkit==0.11.6
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
wrapt==1.14.1
wsproto==1.2.0
